requirements.txt
amqp==2.5.2
anyjson==0.3.3
billiard==3.6.1.0
black-goat-client==0.2.3
celery==4.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.7
django-braces==1.13.0
django-celery==3.3.1
django-cors-headers==3.1.1
django-filter==2.2.0
django-guardian==2.1.0
djangorestframework==3.10.3
funcsigs==1.0.2
gunicorn==19.9.0
jsonpickle==1.2
kombu==4.6.5
lxml==4.4.1
mock==3.0.5
oauthlib==3.1.0
pbr==5.4.3
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.13.0
static3==0.7.0
vine==1.3.0
whitenoise==4.1.4
